+++ Views/FeaturesPage.xaml.cs
@@ -3,13 +3,15 @@ using CommunityToolkit.WinUI.Controls;
 using Microsoft.UI.Xaml;
 using Microsoft.UI.Xaml.Controls;
 using Microsoft.UI.Xaml.Media;
+using Microsoft.Win32;
 using RyTuneX.Helpers;
-using Windows.Storage;
 
 namespace RyTuneX.Views;
 
 public sealed partial class FeaturesPage : Page
 {
+    private const string RegistryBaseKey = @"SOFTWARE\RyTuneX\Optimizations";
+
     public FeaturesPage()
     {
         InitializeComponent();
@@ -17,27 +19,30 @@ public sealed partial class FeaturesPage : Page
         this.NavigationCacheMode = Microsoft.UI.Xaml.Navigation.NavigationCacheMode.Required;
         Loaded += (sender, e) => InitializeToggleSwitchesAsync();
     }
+
     private async void InitializeToggleSwitchesAsync()
     {
         await LogHelper.Log("Initializing Toggle Switches");
         try
         {
-            var tasks = FindVisualChildren<ToggleSwitch>(this).Select(async control =>
+            foreach (var toggleSwitch in FindVisualChildren<ToggleSwitch>(this))
             {
-                if (control.Tag != null && control.Tag is string tagName)
+                if (toggleSwitch.Tag is string tagName)
                 {
-                    // Set the initial state based on the stored value in LocalSettings
-                    var settingValueObj = ApplicationData.Current.LocalSettings.Values[tagName];
-
-                    if (settingValueObj != null && settingValueObj is bool settingValue)
+                    // Retrieve the state from the 64-bit registry with 32-bit app
+                    using var key = RegistryKey.OpenBaseKey(RegistryHive.LocalMachine,
+                        Environment.Is64BitOperatingSystem && !Environment.Is64BitProcess
+                            ? RegistryView.Registry64
+                            : RegistryView.Default).CreateSubKey(RegistryBaseKey);
+                    if (key != null && key.GetValue(tagName) is int state)
                     {
-                        // Subscribe to the Toggled event
-                        control.IsOn = settingValue;
+                        toggleSwitch.IsOn = state == 1;
                     }
-                    control.Toggled += ToggleSwitch_Toggled;
+
+                    // Subscribe to the Toggled event
+                    toggleSwitch.Toggled += ToggleSwitch_Toggled;
                 }
-            });
-            await Task.WhenAll(tasks);
+            }
         }
         catch (Exception ex)
         {
@@ -79,12 +84,19 @@ public sealed partial class FeaturesPage : Page
         {
             var toggleSwitch = (ToggleSwitch)sender;
             Debug.WriteLine($"ToggleSwitch Tag: {toggleSwitch.Tag}, IsOn: {toggleSwitch.IsOn}");
+
+            // Save the state to the 64-bit registry with 32-bit app
+            using var key = RegistryKey.OpenBaseKey(RegistryHive.LocalMachine,
+                Environment.Is64BitOperatingSystem && !Environment.Is64BitProcess
+                    ? RegistryView.Registry64
+                    : RegistryView.Default).CreateSubKey(RegistryBaseKey);
+            key?.SetValue((string)toggleSwitch.Tag, toggleSwitch.IsOn ? 1 : 0, RegistryValueKind.DWord);
+
             await OptimizationOptions.XamlSwitchesAsync(toggleSwitch);
-            ApplicationData.Current.LocalSettings.Values[(string)toggleSwitch.Tag] = toggleSwitch.IsOn;
         }
         catch (Exception ex)
         {
-            await LogHelper.ShowErrorMessageAndLog(ex, XamlRoot);
+            await LogHelper.LogError(ex.Message);
         }
     }
 }